+++ src/ResourceManager/Network/Commands.Network/Cortex/VpnConnection/NewAzureRmVpnConnectionCommand.cs
@@ -159,6 +159,11 @@ namespace Microsoft.Azure.Commands.Network
 
         [Parameter(
             Mandatory = false,
+            HelpMessage = "Enable internet security for this connection")]
+        public SwitchParameter EnableInternetSecurity { get; set; }
+
+        [Parameter(
+            Mandatory = false,
             HelpMessage = "Run cmdlet in the background")]
         public SwitchParameter AsJob { get; set; }
 
@@ -214,6 +219,7 @@ namespace Microsoft.Azure.Commands.Network
             {
                 Name = this.Name,
                 EnableBgp = this.EnableBgp.IsPresent,
+                EnableInternetSecurity = this.EnableInternetSecurity.IsPresent
             };
 
             //// Resolve the VpnSite reference